UI Improvements (#6)
* Move the menu button to the right

This makes it consistent with most other csd apps. Calculator, Fonts, Disks,
Glade, to name a few.

* Remove the unneeded label

Besides the label and separator not looking very good, a user hits the add
or edit button to get to these pages and knows where they are. Instead of
the label make use of the headerbar subtitle if we want an additional hint.
That's part of why it exists.

* Properly center the buttons and error label

The buttons not being ceneterd is most obvious on the edit page. Keep them
centered on the grid.

* Use a wider entry by default

It's nice to be able to see a bit more of what you typed in

* Center the buttons in the main views inline toolbar