Only import bs4 when actually needed
Users reported this issue:

	AttributeError: module 'html5lib.treebuilders' has no attribute '_base'

Making webapp-manager crash on launch.

It's unclear what the issue actually is and whether it's introduced by using PIP
in Mint or by repository versions themselves, but it did affect a few people.

This commit doesn't solve the issue but it confines it to just the icon download
fallback, which itself is within a try/catch, so it no longer makes the app
crash.