Improve icon detection
Only return the results from Favicon Grabber if they are not empty and
the request succeeded.

Scan for all icons with a given format, as they can come in different
sizes, particularly those defined as apple-touch-icon.

Fix msapplication-TileImage finding, which is encoded in a meta tag
rather tan a rel one. See:

https://docs.microsoft.com/en-us/previous-versions/windows/internet-explorer/ie-developer/platform-apis/dn255024(v=vs.85)

Add a few other Microsoft formats, as per the same page.

Redactor and simplify handling slightly by defining parsers for each
type of URL.