Detect Brave Browser for RPM builds (Fedora)
In Fedora Brave Browser binary is saved as `/usr/bin/brave-browser` instead of `/usr/bin/brave`. 
This causes Brave to be unavailable in to not list it in the browser selection drop-down menu. 
This could be solved by creating symbolic link using the following command:
```
$ ln /usr/bin/brave-browser /usr/bin/brave
```
This PR is to add support for `brave-browser` binaries without the user having to create symbolic links.